Create Evaluate Reverse Polish Notation
Given an array of strings tokens that represents an arithmetic expression in a Reverse Polish Notation.
Given Conditions
Evaluate the expression. Return an integer that represents the value of the expression.
The valid operators are '+', '-', '*', and '/'.
Each operand may be an integer or another expression.
The division between two integers always truncates toward zero.
There will not be any division by zero.
The input represents a valid arithmetic expression in a reverse polish notation.